LPS-153267 Splitting item actions into separate component for dropdown and quick actions
